+++ noodles-cram/src/reader/record.rs
@@ -8,9 +8,7 @@ use bytes::Buf;
 use noodles_bam as bam;
 use noodles_core::Position;
 use noodles_sam::{
-    self as sam,
-    record::{quality_scores::Score, sequence::Base},
-    AlignmentRecord,
+    self as sam, alignment::record::sequence::Base, record::quality_scores::Score, AlignmentRecord,
 };
 
 use super::num::get_itf8;