+++ noodles-cram/src/record/features/with_positions.rs
@@ -93,7 +93,7 @@ mod tests {
     #[test]
     fn test_next() -> Result<(), Box<dyn std::error::Error>> {
         use crate::record::Features;
-        use noodles_sam::record::{quality_scores::Score, sequence::Base};
+        use noodles_sam::{alignment::record::sequence::Base, record::quality_scores::Score};
 
         let features = Features::from(vec![
             Feature::Bases(Position::MIN, vec![Base::A, Base::C]),