+++ noodles-sam/src/alignment.rs
@@ -1,0 +1,3 @@
+//! Alignment types.
+
+pub mod record;