+++ noodles-sam/src/alignment/record.rs
@@ -1,0 +1,5 @@
+//! Alignment record.
+
+pub mod sequence;
+
+pub use self::sequence::Sequence;